Make intializer transactional
